+++ frontend/src/components/concept/ActivatableConcept.js
