+++ frontend/src/components/course/AltCourse.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import '../App.css'
 
 const AltCourse = (props) => {
   return (
