+++ frontend/src/components/course/AltCourseForm.js
