+++ frontend/src/components/course/AltCourseList.js
@@ -3,7 +3,7 @@ import AltCourse from './AltCourse'
 import AltCourseForm from './AltCourseForm'
 
 import { useQuery, useMutation } from 'react-apollo-hooks'
-import { ALL_COURSES, CREATE_COURSE } from '../services/CourseService'
+import { ALL_COURSES, CREATE_COURSE } from '../../services/CourseService'
 
 const CourseList = () => {
   const courses = useQuery(ALL_COURSES)
