+++ frontend/src/components/course/CourseContainer.js
