+++ frontend/src/components/course/CourseTray.js
