+++ frontend/src/components/resource/Resource.js
